// src/shared/api/orval/types/bookingSeatResponseGrade.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Smarter Store API
 * Smarter Store 백엔드 API 문서

**관리자 대시보드**: [매출 현황 대시보드](/admin/dashboard.html)
 * OpenAPI spec version: 1.0.0
 */

/**
 * 좌석 등급 (VIP: VIP석, R: R석, S: S석, A: A석, B: B석)
 */
export type BookingSeatResponseGrade = typeof BookingSeatResponseGrade[keyof typeof BookingSeatResponseGrade];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BookingSeatResponseGrade = {
  VIP: 'VIP',
  R: 'R',
  S: 'S',
  A: 'A',
  B: 'B',
} as const;